+++ test/utils/snapshot_with_fake_time.ts
@@ -174,7 +174,10 @@ async function executeTest(
     if (options.denoArgs) {
       denoArgs = options.denoArgs
     } else {
-      denoArgs = ["--quiet", "--allow-env=SNAPSHOT_TEST_NAME"]
+      denoArgs = [
+        "--quiet",
+        "--allow-env=SNAPSHOT_TEST_NAME,XDG_CONFIG_HOME,HOME,APPDATA",
+      ]
     }
 
     // Add FakeTime env var permission if needed